feat: add separators prop and deprecate delimeters
